[BUG] seq +2 when the packet passes through pcie_tx_do_amsdu.
Each skb increases seqno by +1 if the packet passes through pcie_tx_xmit.
We see from the log that pcie_tx_xmit_scheduler does not increase seqno.

Being in BA Stream, is it the firmware that increments each packet?
If yes, and if it is not possible to tell it that we are doing AMSDU, and the client receiving the packet does not process the data if the packet has the same seqno, then we should consider not doing AMSDU.