new queue mapping + add level_threshold + debug
if you look at the files /sys/kernel/debug/ieee80211/phy0/mwlwifi/info or /sys/kernel/debug/ieee80211/phy1/mwlwifi/info,
you notice that the memory addresses are the same, so share.
You understand that the hw_queues are also shared.
I found a conf that uses for the 2Ghz band the hw_queue > 3.

To see if it is not the fact of being in AP which unblocks the hw_queue > 3.

Signed-off-by: jbsky <webmaster@jbsky.fr>